Improve functionaly copyButton
