+++ styles/Tvinfo.module.css
@@ -38,26 +38,6 @@
 .title{
     margin-top: 0;
 }
-.copybutton_box {
-    padding-left: 6px;
-}
-.copyicon {
-    color: #757575;
-}
-.copyicon:hover {
-    cursor: pointer;
-    color: #9e9e9e;
-}
-.copyicon:active {
-    color: #bdbdbd;
-}
-
-.iconlabel {
-    margin-top: 0;
-    padding-top: 1px;
-    color: #424242;
-    font-size: smaller;
-}
 
 .overview_header {
     border-bottom: 1px solid;
